Painting!!! And more item can now be activated

+++ src/main/java/cn/nukkit/item/Item.java
@@ -1357,7 +1357,7 @@ public class Item implements Cloneable {
         return 1;
     }
 
-    public boolean onActivate(Level level, Player player, Block block, Block target, double face, double fx, double fy, double fz) {
+    public boolean onActivate(Level level, Player player, Block block, Block target, int face, double fx, double fy, double fz) {
         return false;
     }
